+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - No longer tested on NodeJS &lt; 16.3.0 and NPM &lt; 7.15.1
 - No longer tested on Ubuntu 16.04 (tested on Ubuntu 18.04)
 - OpenVeo Portal is no longer tested on Opera and Edge
+- Command `grunt build:admin` has been replaced by `npm run build:back-office-client`
 
 # 7.0.0 / 2020-11-19